requirements.txt
# OCR & Image Processing
pytesseract==0.3.13
Pillow==12.1.0
numpy==2.2.6

# Data Processing
pandas==2.3.3

# Web Interface
streamlit==1.52.2

# Utilities
tqdm==4.67.1

# Document Export
python-docx==1.2.0
